fix(position/rotation/scale): fix invalid int->float attr crash, add tests, closes #224, #208

+++ arena/attributes/position.py
@@ -23,7 +23,7 @@ class Position(Attribute):
                 raise ValueError("Position takes x,y,z; a 3-element array or list; or a dict with {x,y,z}")
         elif isinstance(x, str):
             if y is None:
-                x, y, z = map(int, x.split())
+                x, y, z = map(float, x.split())
             else:
                 raise ValueError("Position takes x,y,z; a 3-element array or list; or a dict with {x,y,z}")
         elif isinstance(x, Iterable):
